ImageProcess_Service/requirements.txt
# Image Processing Service Dependencies
# Python 3.10+ required

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image processing
Pillow>=10.0.0

# CLI interface
typer[all]>=0.9.0
rich>=13.0.0

# Async support
asyncio-throttle>=1.0.0

# HTTP client (for testing)
httpx>=0.25.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0